(PC-22875)[API] refactor: rename zendesk_sell backend
