+++ api/.env.development
@@ -96,4 +96,4 @@ WEBAPP_V2_REDIRECT_URL=http://localhost:3000
 CORS_ALLOWED_ORIGINS_BACKOFFICE=http://localhost:.*
 USE_FAST_AND_INSECURE_PASSWORD_HASHING_ALGORITHM=1
 ZENDESK_SELL_API_URL=https://api.getbase.com
-ZENDESK_SELL_BACKEND=pcapi.core.external.backends.logger.LoggerBackend
+ZENDESK_SELL_BACKEND=pcapi.core.external.zendesk_sell_backends.logger.LoggerBackend
