+++ api/.env.integration
@@ -76,4 +76,4 @@ SMS_NOTIFICATION_BACKEND=pcapi.notifications.sms.backends.sendinblue.SendinblueB
 WEBAPP_V2_URL=https://integration.passculture.app
 WEBAPP_V2_REDIRECT_URL=https://redirect.integration.passculture.app
 ZENDESK_SELL_API_URL=https://api.getbase.com
-ZENDESK_SELL_BACKEND=pcapi.core.external.backends.logger.LoggerBackend
+ZENDESK_SELL_BACKEND=pcapi.core.external.zendesk_sell_backends.logger.LoggerBackend
