+++ api/.env.production
@@ -101,4 +101,4 @@ WEBAPP_V2_URL=https://passculture.app
 WEBAPP_V2_REDIRECT_URL=https://redirect.passculture.app
 ZENDESK_API_URL=https://passculture.zendesk.com/api/v2
 ZENDESK_SELL_API_URL=https://api.getbase.com
-ZENDESK_SELL_BACKEND=pcapi.core.external.backends.zendesk_sell.ZendeskSellBackend
+ZENDESK_SELL_BACKEND=pcapi.core.external.zendesk_sell_backends.zendesk_sell.ZendeskSellBackend
