+++ api/.env.staging
@@ -98,5 +98,5 @@ UBBLE_MOCK_CONFIG_URL=https://mock-ubble.ehp.passculture.team
 WEBAPP_V2_URL=https://app.staging.passculture.team
 WEBAPP_V2_REDIRECT_URL=https://redirect.staging.passculture.team
 ZENDESK_SELL_API_URL=https://api.getbase.com
-ZENDESK_SELL_BACKEND=pcapi.core.external.backends.zendesk_sell.ZendeskSellReadOnlyBackend
+ZENDESK_SELL_BACKEND=pcapi.core.external.zendesk_sell_backends.zendesk_sell.ZendeskSellReadOnlyBackend
 BIG_QUERY_NOTIFICATIONS_TABLE_BASENAME="passculture-data-ehp.backend_stg"
