+++ api/.env.testauto
@@ -31,7 +31,7 @@ UBBLE_WEBHOOK_SECRET=fake_ubblewebhook_secret # ggignore
 UBBLE_MOCK_API_URL=
 WEBAPP_V2_REDIRECT_URL=https://webapp-v2.example.com
 WEBAPP_V2_URL=https://webapp-v2.example.com
-ZENDESK_SELL_BACKEND=pcapi.core.external.backends.testing.TestingBackend
+ZENDESK_SELL_BACKEND=pcapi.core.external.zendesk_sell_backends.testing.TestingBackend
 EMS_API_URL=https://fake_url.com
 EMS_API_USER=fake_user
 EMS_API_PASSWORD=fake_password
