+++ api/.env.testing
@@ -101,6 +101,6 @@ UBBLE_MOCK_CONFIG_URL=https://mock-ubble.ehp.passculture.team
 WEBAPP_V2_URL=https://app.testing.passculture.team
 WEBAPP_V2_REDIRECT_URL=https://redirect.testing.passculture.team
 ZENDESK_SELL_API_URL=https://api.getbase.com
-ZENDESK_SELL_BACKEND=pcapi.core.external.backends.logger.LoggerBackend
+ZENDESK_SELL_BACKEND=pcapi.core.external.zendesk_sell_backends.logger.LoggerBackend
 BIG_QUERY_NOTIFICATIONS_TABLE_BASENAME="passculture-data-ehp.backend_dev"
 SESSION_COOKIE_SAMESITE="None"
